Add support for operating on byte strings
This adds a `bytes` submodule for operating on byte strings that might
contain invalid UTF-8. Where possible I have switched the functions that
operate on `str` to use the `bytes` functions internally to avoid
duplicating code and eliminate the potential for differing behavior
between the two functions.

It includes trivial tests that confirm that the `bytes` version of
functions actually work on invalid UTF-8.

Fixes #12